Cast unit_multiplier to float before insert

+++ dags/etl_order_items_initial_load.py
@@ -67,7 +67,7 @@ def _create_initial_order_items_table(ti, xcom_name, truncate=True):
     df["substitute_type"] = df["substitute_type"].astype("int", errors="ignore")
     df["brand"] = df["brand"].astype("int", errors="ignore")
     df["category"] = df["category"].astype("int", errors="ignore")
-    df["unit_multiplier"] = df["unit_multiplier"].astype("int", errors="ignore")
+    df["unit_multiplier"] = df["unit_multiplier"].astype("float", errors="ignore")
 
     columns_rename = {
         "seq_id": "id_orden",
